% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_count.R
\name{binary_count}
\alias{binary_count}
\title{Count the ones in a number in binary}
\usage{
binary_count(num)
}
\arguments{
\item{num}{An integer.}
}
\value{
The result of how many "1"s in the number when it is switched to binary
}
\description{
Count the ones in a number in binary
}
\examples{
binary_count(5)
}
